python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "optiqal"
version = "0.1.0"
description = "Bayesian QALY estimation for lifestyle interventions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Max Ghenis" }
]
keywords = ["qaly", "health", "bayesian", "mcmc", "lifestyle"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "scipy>=1.10",
]

[project.optional-dependencies]
bayesian = [
    "pymc>=5.10",
    "arviz>=0.17",
    "xarray>=2023.1",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = ["optiqal[bayesian,dev]"]

[project.urls]
Homepage = "https://optiqal.ai"
Repository = "https://github.com/PolicyEngine/optiqal-ai"

[tool.hatch.build.targets.wheel]
packages = ["optiqal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
